Approve / reject news broadcast function (close #1)

+++ module/scraper_notices.py
@@ -4,6 +4,9 @@ import ast
 import os
 import copy
 import yaml
+import time
+
+from telegram import InlineKeyboardButton, InlineKeyboardMarkup
 
 with open('config/settings.yaml', 'r') as yaml_config:
     config_map = yaml.load(yaml_config)
@@ -63,7 +66,6 @@ def get_notice_content(notice_dict, base_url, archive_p, notice_p):
     with open(notice_p, 'w') as fw:
         fw.write(formatted_notice)
 
-
 def spam_news(bot, notice_p, channel):
     if os.path.isfile(notice_p):
         message = open(notice_p).read()
@@ -72,6 +74,52 @@ def spam_news(bot, notice_p, channel):
                 bot.sendMessage(chat_id=channel, text=message, parse_mode='HTML')
             except Exception as error:
                 open("logs/errors.txt", "a+").write("{} {}\n".format(error, channel))
+        os.remove(notice_p)
+
+# broadcasts a news passed as a direct message in parameters
+def spam_news_direct(bot, notice_message, channel):
+    if notice_message != "":
+        try:
+            bot.sendMessage(chat_id=channel, text=notice_message, parse_mode='HTML')
+        except Exception as error:
+            open("logs/errors.txt", "a+").write("{} {}\n".format(error, channel))
+
+def send_news_approve_message(bot, notice_p, channel_folder, channel, group_chatid):
+    # maybe pending approval folder should be settable, to be reviewed
+    pending_approval_folder = "in_approvazione"
+
+    if os.path.isfile(notice_p):
+        notice_message = open(notice_p).read()
+
+        if notice_message != "":
+            try:
+                # notice disk id is used to identify an approval pending message. OS clock's used for this
+                notice_disk_id = time.clock()
+                approving_notice_filename = "{}/{}/{}_{}.dat".format(channel_folder, pending_approval_folder, channel, notice_disk_id)
+
+                if not os.path.exists(os.path.dirname(approving_notice_filename)):
+                    try:
+                        os.makedirs(os.path.dirname(approving_notice_filename))
+                    except OSError as exc: 
+                        if exc.errno != errno.EEXIST:
+                            raise
+
+                # write notice data into a file
+                with open(approving_notice_filename, 'w') as fw:
+                    fw.writelines(notice_message[0:])
+
+                # reply buttons layout
+                keyboard_markup = [
+                    [InlineKeyboardButton("Accetta ✔", callback_data="news:approved:{}:{}:{}".format(channel, channel_folder, notice_disk_id)),
+                    InlineKeyboardButton("Rifiuta ❌", callback_data="news:rejected:{}:{}:{}".format(channel, channel_folder, notice_disk_id))]
+                ]
+
+                reply_markup = InlineKeyboardMarkup(keyboard_markup)
+
+                # finally, send the message to the approval group
+                bot.sendMessage(chat_id=group_chatid, text=notice_message, parse_mode='HTML', reply_markup=reply_markup)
+            except Exception as error:
+                open("logs/errors.txt", "a+").write("send_news_approve_message: {} {}\n".format(error, channel))
         os.remove(notice_p)
 
 
@@ -87,8 +135,8 @@ def scrape_notices(bot, job):
         else:
             folder = i
 
-        pending_path = "data/avvisi/"+str(folder)+"/"+str(ch)+"_avvisi.dat"
-        archive_path = "data/avvisi/"+str(folder)+"/"+str(ch)+"_avvisi_in_sospeso.dat"
+        pending_path = "data/avvisi/"+str(folder)+"/"+str(ch)+"_avvisi_in_sospeso.dat"
+        archive_path = "data/avvisi/"+str(folder)+"/"+str(ch)+"_avvisi.dat"
         notice_path = "data/avvisi/"+str(folder)+"/"+str(ch)+"_avviso.dat"
 
         base_url = notices_urls_cp[i]["urls"][list(notices_urls_cp[i]["urls"])[0]]
@@ -122,5 +170,10 @@ def scrape_notices(bot, job):
             pending_notice = pull_pending_notice(pending_path)
             if pending_notice:
                 get_notice_content(pending_notice, base_url, archive_path, notice_path)
-        
-        spam_news(bot, notice_path, notices_urls[i]["channel"])
+
+        approve_group_chatid = notices_urls[i]["approve_group_chatid"]
+
+        if approve_group_chatid:
+            send_news_approve_message(bot, notice_path, "data/avvisi/"+str(folder), notices_urls[i]["channel"], approve_group_chatid)
+        else:
+            spam_news(bot, notice_path, notices_urls[i]["channel"])
